Add tags columns missing
When we dump the using osmosis we also get from the osm files
the tags field (that will be hstore field in table). So we add it to
the schema defs. We should properly handle this case with sqitch
but at this time we are adding it directly to the sqitch deploy files